Moved snapshot provisioning behind provider hooks
